requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-SocketIO==5.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
Werkzeug==2.0.1